Create a Pre-defined Directory Hierarchy on Disk

Create a Pre-defined Directory Hierarchy on Disk

Usage

create_directory_structure(
  path = getwd(),
  structure_path = system.file("templates", "_project_structure_en.yaml", package =
    "saros.utils"),
  numbering_prefix = c("none", "max_local", "max_global"),
  numbering_inheritance = TRUE,
  word_separator = NULL,
  numbering_parent_child_separator = word_separator,
  numbering_name_separator = " ",
  case = c("asis", "sentence", "title", "lower", "upper", "snake"),
  replacement_list = c(project_initials = "SSN"),
  create = FALSE,
  count_existing_folders = FALSE
)

Arguments

path

String, path to where to create the project files

structure_path

String. Path to the YAML file that defines the folder structure. Defaults to system.file("templates", "_project_structure_en.yaml").

numbering_prefix

String. One of c("none", "max_local", "max_global").

numbering_inheritance

Flag. Whether to inherit numbering from parent folder.

word_separator

String. Replace separators between words in folder names. Defaults to NULL.

numbering_parent_child_separator

String. Defaults to word_separator.

numbering_name_separator

String. Separator between numbering part and name.

case

String. One of c("asis", "sentence", "lower", "upper", "title", "snake").

replacement_list

named character vector. Each name in this vector will be replaced with its "{{value}}" in the structure_path file

create

Boolean. Defaults to TRUE in initialize_saros_project(), FALSE in create_directory_structure().

count_existing_folders

Boolean. Defaults to FALSE.

Value

No return value, called for side effects

Examples

create_directory_structure()
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application/Call 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application/Formalities 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application/CVs 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application/Collaboration and cooperation agreements 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application/Application 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application/Pre-analysis 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Application/For submission 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Budget 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Contracts and agreements 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Invoices, accounting and receipts 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Status reports 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Logo and graphical materials 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Internal meetings 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Internal meetings/Meeting agendas and invitations 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Internal meetings/Internal presentations 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Administration/Internal meetings/Minutes 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Overall planning 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Consent form 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Ethical-GDPR approval 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Survey questionnaires 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Staff 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Pupils 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Parents 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Researchers 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Leaders 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Teachers 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Principals 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Students 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Interview guides/Population 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Request of data from 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Literature review-design 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Intervention materials 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Randomizing participants 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Materials/Chapter overviews 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Literature 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Literature/Topic has policy relevance 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Literature/Pure theory and framework 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Literature/Similar empirical studies 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Literature/Similar instruments and guides for data collection 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Literature/Relevant analytic methodology 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Literature/Unprocessed (remove from here) 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Population data 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Population data/Codebook 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Sampling frame 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Registry data 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Collected respondent lists 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Respondent list for survey system 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Downloaded response data 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Downloaded response data/Codebook 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Qualitative data 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Qualitative data/Interview recordings 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Qualitative data/Observational notes 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Text corpus 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/PDF-reports 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Prepared data 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Data/Prepared data/Codebooks 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Scripts 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Resources 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Draft generations 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Draft generations/main 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Draft generations/Reports 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Drafts in editing 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Drafts in editing/main 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Drafts in editing/Reports 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Completed drafts 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Completed drafts/main 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Saros_SSN/Completed drafts/Reports 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials) 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials)/Cover letter & response to reviewers 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials)/Manuscript 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials)/Analysis 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials)/Figures for submission 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials)/Tables for submission 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials)/Appendix 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Publications/Paper1-Short title (author initials)/Online supplementary materials 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Outreach 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Outreach/Research conference presentation 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Outreach/Research conference poster 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Outreach/Stakeholders and reference group 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Outreach/Stakeholders' communication channels 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Outreach/Practitioners and special interest channels 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Outreach/Public through mass media channels 
#> /home/runner/work/saros.utils/saros.utils/docs/reference/Other