Post-render rendering of complete reports

If one wishes to render complete reports that are linked to in a website, but not listed among the chapters on the sidebar menu, one can make these with filenames starting with underscores (_) which will make them not listed in said menu. However, this will also mean they are not rendered in a Quarto (Website) project. This function, which can be called within a post-script (see example below), will render only these after the ordinary rendering of the project, and copied to the _site folder.

Usage

render_full_reports(
  files = NULL,
  path,
  processable_path = file.path(path, "Reports"),
  site_path = file.path(path, "_site"),
  resource_paths = file.path(path, c("_extensions", "_images")),
  warn_on_file_error = FALSE,
  ...
)

Arguments

files

Optional character vector of report files (qmd). Can be obtained within a project by Sys.getenv("QUARTO_PROJECT_OUTPUT_FILES")

path

If no files are given, a path to the root folder of the local "site".

processable_path

Path to where report files can be (recursively) found.

site_path

Path to _site

resource_paths

Paths to where _extensions and _images folders can be found and copied to wherever needed

warn_on_file_error

If TRUE, will collect warnings if a file fails to render or be copied. If FALSE (default), will stop the rendering process.

...

Additional arguments passed to quarto::render()

Value

Returns invisibly a character vector of processed files.